Module osmosis_std::types::cosmos::auth::v1beta1

Structs

BaseAccount

BaseAccount defines a base account type. It contains all the necessary fields for basic account functionality. Any custom account type should extend this type for additional functionality (e.g. vesting).

ModuleAccount

ModuleAccount defines an account for modules that holds coins on a pool.

Params

Params defines the parameters for the auth module.