Module osmosis_std::types::osmosis::gamm::v1beta1

Structs

GammQuerier

GenesisState

GenesisState defines the gamm module’s genesis state.

MsgExitPool

===================== MsgExitPool

MsgExitPoolResponse

MsgExitSwapExternAmountOut

===================== MsgExitSwapExternAmountOut

MsgExitSwapExternAmountOutResponse

MsgExitSwapShareAmountIn

===================== MsgExitSwapShareAmountIn

MsgExitSwapShareAmountInResponse

MsgJoinPool

===================== MsgJoinPool This is really MsgJoinPoolNoSwap

MsgJoinPoolResponse

MsgJoinSwapExternAmountIn

===================== MsgJoinSwapExternAmountIn TODO: Rename to MsgJoinSwapExactAmountIn

MsgJoinSwapExternAmountInResponse

MsgJoinSwapShareAmountOut

===================== MsgJoinSwapShareAmountOut

MsgJoinSwapShareAmountOutResponse

MsgSwapExactAmountIn

MsgSwapExactAmountInResponse

MsgSwapExactAmountOut

MsgSwapExactAmountOutResponse

Params

Params holds parameters for the incentives module

Pool

PoolAsset

Pool asset is an internal struct that combines the amount of the token in the pool, and its balancer weight. This is an awkward packaging of data, and should be revisited in a future state migration.

PoolParams

PoolParams defined the parameters that will be managed by the pool governance in the future. This params are not managed by the chain governance. Instead they will be managed by the token holders of the pool. The pool’s token holders are specified in future_pool_governor.

QueryNumPoolsRequest

=============================== NumPools

QueryNumPoolsResponse

QueryPoolParamsRequest

=============================== PoolParams

QueryPoolParamsResponse

QueryPoolRequest

=============================== Pool

QueryPoolResponse

QueryPoolsRequest

=============================== Pools

QueryPoolsResponse

QuerySpotPriceRequest

QuerySpotPriceRequest defines the gRPC request structure for a SpotPrice query.

QuerySpotPriceResponse

QuerySpotPriceResponse defines the gRPC response structure for a SpotPrice query.

QuerySwapExactAmountInRequest

=============================== EstimateSwapExactAmountIn

QuerySwapExactAmountInResponse

QuerySwapExactAmountOutRequest

=============================== EstimateSwapExactAmountOut

QuerySwapExactAmountOutResponse

QueryTotalLiquidityRequest

QueryTotalLiquidityResponse

QueryTotalPoolLiquidityRequest

=============================== PoolLiquidity

QueryTotalPoolLiquidityResponse

QueryTotalSharesRequest

=============================== TotalShares

QueryTotalSharesResponse

SmoothWeightChangeParams

Parameters for changing the weights in a balancer pool smoothly from a start weight and end weight over a period of time. Currently, the only smooth change supported is linear changing between the two weights, but more types may be added in the future. When these parameters are set, the weight w(t) for pool time t is the following: t <= start_time: w(t) = initial_pool_weights start_time < t <= start_time + duration: w(t) = initial_pool_weights + (t - start_time) * (target_pool_weights - initial_pool_weights) / (duration) t > start_time + duration: w(t) = target_pool_weights

SwapAmountInRoute

===================== MsgSwapExactAmountIn

SwapAmountOutRoute

===================== MsgSwapExactAmountOut