Module osmosis_std::types::osmosis::incentives

Structs

ActiveGaugesPerDenomRequest

ActiveGaugesPerDenomResponse

ActiveGaugesRequest

ActiveGaugesResponse

Gauge

Gauge is an object that stores and distributes yields to recipients who satisfy certain conditions. Currently gauges support conditions around the duration for which a given denom is locked.

GaugeByIdRequest

GaugeByIdResponse

GaugesRequest

GaugesResponse

GenesisState

GenesisState defines the incentives module’s various parameters when first initialized

IncentivesQuerier

LockableDurationsInfo

ModuleDistributedCoinsRequest

ModuleDistributedCoinsResponse

ModuleToDistributeCoinsRequest

ModuleToDistributeCoinsResponse

MsgAddToGauge

MsgAddToGauge adds coins to a previously created gauge

MsgAddToGaugeResponse

MsgCreateGauge

MsgCreateGauge creates a gague to distribute rewards to users

MsgCreateGaugeResponse

Params

Params holds parameters for the incentives module

QueryLockableDurationsRequest

QueryLockableDurationsResponse

RewardsEstRequest

RewardsEstResponse

UpcomingGaugesPerDenomRequest

UpcomingGaugesPerDenomResponse

UpcomingGaugesRequest

UpcomingGaugesResponse