Module osmosis_std::types::osmosis::lockup

Structs

AccountLockedCoinsRequest

AccountLockedCoinsResponse

AccountLockedDurationRequest

AccountLockedDurationResponse

AccountLockedLongerDurationDenomRequest

AccountLockedLongerDurationDenomResponse

AccountLockedLongerDurationNotUnlockingOnlyRequest

AccountLockedLongerDurationNotUnlockingOnlyResponse

AccountLockedLongerDurationRequest

AccountLockedLongerDurationResponse

AccountLockedPastTimeDenomRequest

AccountLockedPastTimeDenomResponse

AccountLockedPastTimeNotUnlockingOnlyRequest

AccountLockedPastTimeNotUnlockingOnlyResponse

AccountLockedPastTimeRequest

AccountLockedPastTimeResponse

AccountUnlockableCoinsRequest

AccountUnlockableCoinsResponse

AccountUnlockedBeforeTimeRequest

AccountUnlockedBeforeTimeResponse

AccountUnlockingCoinsRequest

AccountUnlockingCoinsResponse

GenesisState

GenesisState defines the lockup module’s genesis state.

LockedDenomRequest

LockedDenomResponse

LockedRequest

LockedResponse

LockupQuerier

ModuleBalanceRequest

ModuleBalanceResponse

ModuleLockedAmountRequest

ModuleLockedAmountResponse

MsgBeginUnlocking

MsgBeginUnlockingAll

MsgBeginUnlockingAllResponse

MsgBeginUnlockingResponse

MsgExtendLockup

MsgExtendLockup extends the existing lockup’s duration. The new duration is longer than the original.

MsgExtendLockupResponse

MsgLockTokens

MsgLockTokensResponse

PeriodLock

PeriodLock is a single lock unit by period defined by the x/lockup module. It’s a record of a locked coin at a specific time. It stores owner, duration, unlock time and the number of coins locked. A state of a period lock is created upon lock creation, and deleted once the lock has been matured after the duration has passed since unbonding started.

QueryCondition

QueryCondition is a struct used for querying locks upon different conditions. Duration field and timestamp fields could be optional, depending on the LockQueryType.

SyntheticLock

SyntheticLock is creating virtual lockup where new denom is combination of original denom and synthetic suffix. At the time of synthetic lockup creation and deletion, accumulation store is also being updated and on querier side, they can query as freely as native lockup.

SyntheticLockupsByLockupIdRequest

SyntheticLockupsByLockupIdResponse

Enums

LockQueryType

LockQueryType defines the type of the lock query that can either be by duration or start time of the lock.