Module osmosis_std::types::osmosis::mint::v1beta1

Structs

DistributionProportions

DistributionProportions defines the distribution proportions of the minted denom. In other words, defines which stakeholders will receive the minted denoms and how much.

GenesisState

GenesisState defines the mint module’s genesis state.

MintQuerier

Minter

Minter represents the minting state.

Params

Params holds parameters for the x/mint module.

QueryEpochProvisionsRequest

QueryEpochProvisionsRequest is the request type for the Query/EpochProvisions RPC method.

QueryEpochProvisionsResponse

QueryEpochProvisionsResponse is the response type for the Query/EpochProvisions RPC method.

QueryParamsRequest

QueryParamsRequest is the request type for the Query/Params RPC method.

QueryParamsResponse

QueryParamsResponse is the response type for the Query/Params RPC method.

WeightedAddress

WeightedAddress represents an address with a weight assigned to it. The weight is used to determine the proportion of the total minted tokens to be minted to the address.