Module osmosis_std::types::osmosis::poolincentives::v1beta1

Modules

query_gauge_ids_response

Nested message and enum types in QueryGaugeIdsResponse.

Structs

DistrInfo

DistrRecord

GenesisState

GenesisState defines the pool incentives module’s genesis state.

IncentivizedPool

LockableDurationsInfo

Params

PoolToGauge

PoolToGauges

PoolincentivesQuerier

QueryDistrInfoRequest

QueryDistrInfoResponse

QueryExternalIncentiveGaugesRequest

QueryExternalIncentiveGaugesResponse

QueryGaugeIdsRequest

QueryGaugeIdsResponse

QueryIncentivizedPoolsRequest

QueryIncentivizedPoolsResponse

QueryLockableDurationsRequest

QueryLockableDurationsResponse

QueryParamsRequest

QueryParamsResponse

ReplacePoolIncentivesProposal

ReplacePoolIncentivesProposal is a gov Content type for updating the pool incentives. If a ReplacePoolIncentivesProposal passes, the proposal’s records override the existing DistrRecords set in the module. Each record has a specified gauge id and weight, and the incentives are distributed to each gauge according to weight/total_weight. The incentives are put in the fee pool and it is allocated to gauges and community pool by the DistrRecords configuration. Note that gaugeId=0 represents the community pool.

UpdatePoolIncentivesProposal

For example: if the existing DistrRecords were: [(Gauge 0, 5), (Gauge 1, 6), (Gauge 2, 6)] An UpdatePoolIncentivesProposal includes [(Gauge 1, 0), (Gauge 2, 4), (Gauge 3, 10)] This would delete Gauge 1, Edit Gauge 2, and Add Gauge 3 The result DistrRecords in state would be: [(Gauge 0, 5), (Gauge 2, 4), (Gauge 3, 10)]