Module osmosis_std::types::osmosis::superfluid

Modules

v1beta1

Structs

AllAssetsRequest

AllAssetsResponse

AllIntermediaryAccountsRequest

AllIntermediaryAccountsResponse

AssetMultiplierRequest

AssetMultiplierResponse

AssetTypeRequest

AssetTypeResponse

ConnectedIntermediaryAccountRequest

ConnectedIntermediaryAccountResponse

EstimateSuperfluidDelegatedAmountByValidatorDenomRequest

EstimateSuperfluidDelegatedAmountByValidatorDenomResponse

GenesisState

GenesisState defines the module’s genesis state.

LockIdIntermediaryAccountConnection

LockIdIntermediaryAccountConnection is a struct used to indicate the relationship between the underlying lock id and superfluid delegation done via lp shares.

MsgLockAndSuperfluidDelegate

MsgLockAndSuperfluidDelegate locks coins with the unbonding period duration, and then does a superfluid lock from the newly created lockup, to the specified validator addr.

MsgLockAndSuperfluidDelegateResponse

MsgSuperfluidDelegate

MsgSuperfluidDelegateResponse

MsgSuperfluidUnbondLock

MsgSuperfluidUnbondLockResponse

MsgSuperfluidUndelegate

MsgSuperfluidUndelegateResponse

MsgUnPoolWhitelistedPool

MsgUnPoolWhitelistedPool Unpools every lock the sender has, that is associated with pool pool_id. If pool_id is not approved for unpooling by governance, this is a no-op. Unpooling takes the locked gamm shares, and runs “ExitPool” on it, to get the constituent tokens. e.g. z gamm/pool/1 tokens ExitPools into constituent tokens x uatom, y uosmo. Then it creates a new lock for every constituent token, with the duration associated with the lock. If the lock was unbonding, the new lockup durations should be the time left until unbond completion.

MsgUnPoolWhitelistedPoolResponse

OsmoEquivalentMultiplierRecord

The Osmo-Equivalent-Multiplier Record for epoch N refers to the osmo worth we treat an LP share as having, for all of epoch N. Eventually this is intended to be set as the Time-weighted-average-osmo-backing for the entire duration of epoch N-1. (Thereby locking whats in use for epoch N as based on the prior epochs rewards) However for now, this is not the TWAP but instead the spot price at the boundary. For different types of assets in the future, it could change.

Params

Params holds parameters for the superfluid module

QueryParamsRequest

QueryParamsResponse

QueryTotalDelegationByDelegatorRequest

QueryTotalDelegationByDelegatorResponse

SuperfluidAsset

SuperfluidAsset stores the pair of superfluid asset type and denom pair

SuperfluidDelegationAmountRequest

SuperfluidDelegationAmountResponse

SuperfluidDelegationRecord

SuperfluidDelegationRecord is a struct used to indicate superfluid delegations of an account in the state machine in a user friendly form.

SuperfluidDelegationsByDelegatorRequest

SuperfluidDelegationsByDelegatorResponse

SuperfluidDelegationsByValidatorDenomRequest

SuperfluidDelegationsByValidatorDenomResponse

SuperfluidIntermediaryAccount

SuperfluidIntermediaryAccount takes the role of intermediary between LP token and OSMO tokens for superfluid staking. The intermediary account is the actual account responsible for delegation, not the validator account itself.

SuperfluidIntermediaryAccountInfo

SuperfluidQuerier

SuperfluidUndelegationsByDelegatorRequest

SuperfluidUndelegationsByDelegatorResponse

TotalSuperfluidDelegationsRequest

TotalSuperfluidDelegationsResponse

UnpoolWhitelistedPools

Enums

SuperfluidAssetType

SuperfluidAssetType indicates whether the superfluid asset is a native token itself or the lp share of a pool.