Module osmosis_std::types::osmosis::tokenfactory::v1beta1

Structs

DenomAuthorityMetadata

DenomAuthorityMetadata specifies metadata for addresses that have specific capabilities over a token factory denom. Right now there is only one Admin permission, but is planned to be extended to the future.

GenesisDenom

GenesisDenom defines a tokenfactory denom that is defined within genesis state. The structure contains DenomAuthorityMetadata which defines the denom’s admin.

GenesisState

GenesisState defines the tokenfactory module’s genesis state.

MsgBurn

MsgBurn is the sdk.Msg type for allowing an admin account to burn a token. For now, we only support burning from the sender account.

MsgBurnResponse

MsgChangeAdmin

MsgChangeAdmin is the sdk.Msg type for allowing an admin account to reassign adminship of a denom to a new account

MsgChangeAdminResponse

MsgChangeAdminResponse defines the response structure for an executed MsgChangeAdmin message.

MsgCreateDenom

MsgCreateDenom defines the message structure for the CreateDenom gRPC service method. It allows an account to create a new denom. It requires a sender address and a sub denomination. The (sender_address, sub_denomination) tuple must be unique and cannot be re-used.

MsgCreateDenomResponse

MsgCreateDenomResponse is the return value of MsgCreateDenom It returns the full string of the newly created denom

MsgMint

MsgMint is the sdk.Msg type for allowing an admin account to mint more of a token. For now, we only support minting to the sender account

MsgMintResponse

MsgSetDenomMetadata

MsgSetDenomMetadata is the sdk.Msg type for allowing an admin account to set the denom’s bank metadata

MsgSetDenomMetadataResponse

MsgSetDenomMetadataResponse defines the response structure for an executed MsgSetDenomMetadata message.

Params

Params defines the parameters for the tokenfactory module.

QueryDenomAuthorityMetadataRequest

QueryDenomAuthorityMetadataRequest defines the request structure for the DenomAuthorityMetadata gRPC query.

QueryDenomAuthorityMetadataResponse

QueryDenomAuthorityMetadataResponse defines the response structure for the DenomAuthorityMetadata gRPC query.

QueryDenomsFromCreatorRequest

QueryDenomsFromCreatorRequest defines the request structure for the DenomsFromCreator gRPC query.

QueryDenomsFromCreatorResponse

QueryDenomsFromCreatorRequest defines the response structure for the DenomsFromCreator gRPC query.

QueryParamsRequest

QueryParamsRequest is the request type for the Query/Params RPC method.

QueryParamsResponse

QueryParamsResponse is the response type for the Query/Params RPC method.

TokenfactoryQuerier