Module osmosis_std::types::osmosis::txfees::v1beta1

Structs

FeeToken

FeeToken is a struct that specifies a coin denom, and pool ID pair. This marks the token as eligible for use as a tx fee asset in Osmosis. Its price in osmo is derived through looking at the provided pool ID. The pool ID must have osmo as one of its assets.

GenesisState

GenesisState defines the txfees module’s genesis state.

QueryBaseDenomRequest

QueryBaseDenomResponse

QueryDenomPoolIdRequest

QueryDenomPoolIdResponse

QueryDenomSpotPriceRequest

QueryDenomSpotPriceRequest defines grpc request structure for querying spot price for the specified tx fee denom

QueryDenomSpotPriceResponse

QueryDenomSpotPriceRequest defines grpc response structure for querying spot price for the specified tx fee denom

QueryFeeTokensRequest

QueryFeeTokensResponse

TxfeesQuerier

UpdateFeeTokenProposal

UpdateFeeTokenProposal is a gov Content type for adding a new whitelisted fee token. It must specify a denom along with gamm pool ID to use as a spot price calculator. It can be used to add a new denom to the whitelist It can also be used to update the Pool to associate with the denom. If Pool ID is set to 0, it will remove the denom from the whitelisted set.